backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
numpy
scikit-learn
sentence-transformers
torch
requests
